renamed lib/slp/base.py -> lib/slp/slp.py
