+++ lib/slp/__init__.py
@@ -1,1 +1,1 @@
-from .base import *
+from .slp import *
